backend_ai/requirements.txt
absl-py==2.1.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.5.0
anyio==3.7.1
astor==0.8.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
boto3==1.33.13
botocore==1.33.13
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cycler==0.11.0
eth-abi-lite==3.2.0
exceptiongroup==1.2.1
fastapi==0.103.2
fonttools==4.38.0
frozenlist==1.3.3
gast==0.2.2
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
h5py==2.10.0
idna==3.7
importlib-metadata==6.7.0
importlib-resources==5.12.0
jmespath==1.0.1
joblib==1.3.2
jsonpickle==3.0.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
loguru==0.6.0
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdurl==0.1.2
mido==1.3.2
more-itertools==9.1.0
multidict==6.0.5
music21==7.3.3
numpy==1.18.5
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
protobuf==3.20.0
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
PyMuPDF==1.22.5
PyMySQL==1.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-multipart==0.0.8
pytz==2024.1
rich==13.7.1
s3transfer==0.8.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.27.0
tensorboard==1.15.0
tensorflow==1.15.5
tensorflow-estimator==1.15.1
termcolor==2.3.0
toml==0.10.2
toolcache==0.5.0
toolcli==0.6.16
toolconf==0.1.2
toolsql==0.3.11
toolstr==0.8.5
tooltime==0.2.11
typing_extensions==4.7.1
urllib3==1.26.18
uvicorn==0.22.0
webcolors==1.13
Werkzeug==2.2.3
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.15.0
